requirements.txt
telethon==1.41.1
requests==2.32.3
python-dotenv==1.0.1
pytz==2024.1
pytest==8.3.3
apscheduler==3.10.4
aiosqlite==0.20.0
httpx==0.27.0
websockets==12.0
